requirements.txt
openai

#DISCORD
discord.py

#TELEGRAM
python-telegram-bot